examples/basic/python/headers/pyproject.toml
[project]
name = "shapesinc-python-headers"
version = "1.0.0"
description = "An advanced example of using the Shapes API to talk to your shape."
authors = [
    { name = "Shapes Inc.", email = "hi@shapes.inc" }
]
keywords = []
license = { text = "MIT" }
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.8"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.4.26",
    "distro==1.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.9.0",
    "openai==1.76.0",
    "pydantic==2.11.3",
    "pydantic_core==2.33.1",
    "python-dotenv==1.1.0",
    "sniffio==1.3.1",
    "tqdm==4.67.1",
    "typing-inspection==0.4.0",
    "typing_extensions==4.13.2",
]

[project.urls]
Homepage = "https://shapes.inc"
Repository = "https://github.com/shapesinc/api.git"

[project.scripts]
shape-headers = "main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
